More compile fixes from Jason Sundram.
git-svn-id: svn+ssh://svn.code.sf.net/p/marsyas/code/trunk@3227 0e98fd9d-079e-457f-9c5c-d482923f975d

+++ src/marsyas/WekaSource.h
@@ -73,7 +73,7 @@ namespace Marsyas
 			this->Shuffle();
 
 			rstep_ = (mrs_real)this->size() / (mrs_real)foldCount_;
-			if (foldCount_ > this->size()) 
+			if ((unsigned int)foldCount_ > this->size()) 
 			{
 				cout << "Folds exceed number of instances" << endl;
 				cout << "foldCount_ = " << foldCount_ << endl;
